refactor: pagination 공통 로직 유틸함수로 분리 및 setTimeout 제거
